requirementst.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
newsapi-python==0.2.7
redis==5.2.0
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.6